Database level lock file. Database level admin commands.
Before this commit admin commands were trying to check DBMS
lock and it was not possible to execute any individual database
level operation if DBMS was running.
This PR introduces database level lock file in addition to DBMS level
lock that will allow individual commands to be executed on top of
databases and prevent access to the same database from different command
and DBMS processes.

Admin commands were updated to always grab database lock to ensure that
only one process works on top of the database.
It's a slight modification of commands compare to the previous lifecycle
when the whole DBMS lock was either only checked, or totally ignored at all.

This commit also removes custom version checker that was used in dbms module
and update the code to use version classes that are provided by kernel.

+++ community/io/src/main/java/org/neo4j/io/fs/FileUtils.java
@@ -59,6 +59,7 @@ import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.Set;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -68,6 +69,7 @@ import static java.nio.file.StandardOpenOption.READ;
 import static java.nio.file.StandardOpenOption.WRITE;
 import static java.util.Objects.requireNonNull;
 import static org.apache.commons.lang3.reflect.FieldUtils.getDeclaredField;
+import static org.neo4j.function.Predicates.alwaysTrue;
 import static org.neo4j.io.fs.FileSystemAbstraction.INVALID_FILE_DESCRIPTOR;
 import static org.neo4j.util.FeatureToggles.flag;
 
@@ -142,12 +144,27 @@ public class FileUtils
 
     public static void deletePathRecursively( Path path ) throws IOException
     {
+        deletePathRecursively( path, alwaysTrue() );
+    }
+
+    public static void deletePathRecursively( Path path, Predicate<Path> removeFilePredicate ) throws IOException
+    {
+
         Files.walkFileTree( path, new SimpleFileVisitor<>()
         {
+            private int skippedFiles;
+
             @Override
             public FileVisitResult visitFile( Path file, BasicFileAttributes attrs ) throws IOException
             {
-                deleteFile( file );
+                if ( removeFilePredicate.test( file ) )
+                {
+                    deleteFile( file );
+                }
+                else
+                {
+                    skippedFiles++;
+                }
                 return FileVisitResult.CONTINUE;
             }
 
@@ -160,14 +177,22 @@ public class FileUtils
                 }
                 try
                 {
-                    Files.delete( dir );
+                    if ( skippedFiles == 0 )
+                    {
+                        Files.delete( dir );
+                        return FileVisitResult.CONTINUE;
+                    }
+                    if ( Files.list( dir ).noneMatch( alwaysTrue() ) )
+                    {
+                        Files.delete( dir );
+                    }
+                    return FileVisitResult.CONTINUE;
                 }
                 catch ( DirectoryNotEmptyException notEmpty )
                 {
                     String reason = notEmptyReason( dir, notEmpty );
                     throw new IOException( notEmpty.getMessage() + ": " + reason, notEmpty );
                 }
-                return FileVisitResult.CONTINUE;
             }
 
             private String notEmptyReason( Path dir, DirectoryNotEmptyException notEmpty )
